Updating People Page during Fixes
The people page was identified as having an incorrect number of people for the unsheltered graphic. 
Discovered the BNL report was altered and the column for Days since assessment is significantly higher, causing people to be filtered out.
I redid the October Dashboard page, added language to note it is old information, and have uploaded it here.