Add all files to update the dashboard; there are unnecessary files
The dashboard needs to be updated, and I know what I'm pushing works with a local web server (used through node).

The node_modules probably aren't required, but I need to get the site live, so everything is going in just so I can be sure I'm not missing anything.